requirements.txt
scikit-learn
pandas
numpy
pillow
streamlit
opencv-python
matplotlib
joblib
